restyle compare page, fix and modularize popup
