+++ src/components/Button/style.scss
@@ -14,7 +14,6 @@ button {
     border-radius: 5px;
     background: transparent;
     height: 40px;
-    width: 280px;
     z-index: 99;
     outline: none;
 }
