+++ src/components/DuetLogo/index.js
@@ -1,8 +1,9 @@
+/* eslint-disable react/prop-types */
 import React from 'react';
 
-const DuetLogo = () => {
+const DuetLogo = ({ width, height }) => {
     return (
-        <svg width="267" height="162" viewBox="0 0 267 162" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <svg width={width} height={height} viewBox="0 0 267 162" fill="none" xmlns="http://www.w3.org/2000/svg">
             <rect y="85.32" width="266.76" height="76.68" fill="#212034"/>
             <rect x="45.36" width="72.36" height="160.92" fill="#212034"/>
             <path d="M62.3374 145.88C57.5525 152.018 50.0762 155.311 41.7026 155.311C25.8527 155.311 14.04 142.886 14.04 126.269C14.04 109.652 25.8527 97.2267 41.5531 97.2267C49.9266 97.2267 57.2535 100.82 62.1879 106.658V14.142C62.1879 13.3935 62.4869 12.645 62.786 12.0461C63.6832 10.8485 65.1784 10.5491 66.6737 10.9982L107.644 24.9205C109.289 25.6691 109.887 27.6152 108.99 29.2619C108.242 30.9086 106.298 31.5074 104.654 30.6092L69.0662 19.3816V126.269C69.0662 134.802 66.6737 140.191 62.3374 145.88ZM62.3374 126.269C62.3374 113.544 53.3658 103.963 41.7026 103.963C30.0395 103.963 20.9183 113.544 20.9183 126.269C20.9183 138.844 29.89 148.425 41.5531 148.425C53.2162 148.425 62.3374 138.844 62.3374 126.269Z" fill="#E5277B"/>
