--- src/components/InfoAndCompare/Compare/style.scss
@@ -1,87 +1,0 @@
-h4 {
-    color: $SECONDARY;
-}
-
-#content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    z-index: 2;
-    margin-top: 17vh;
-
-    input {
-        border: 1px solid $SECONDARY;
-        border-radius: 5px;
-        background: transparent;
-        height: 30px;
-        width: 280px;
-        color: $SECONDARY
-    }
-
-    button {
-        font-size: 16px;
-        color: $PRIMARY;
-        border-radius: 5px;
-        background: $SECONDARY;
-        height: 30px;
-        width: 80px;
-        z-index: 99;
-        outline:none;
-        margin-top: 30px;
-    }
-
-    #smallLink {
-        color: $SECONDARY;
-        background: transparent;
-        border: none;
-        font-size: 14px;
-        width: auto;
-        margin-top: 10px;
-    }
-}
-
-#PopupBackground {
-    height: 40vh;
-    margin-top: 20vh;
-
-    input {
-        border: 1px solid $SECONDARY;
-        border-radius: 5px;
-        background: transparent;
-        height: 30px;
-        width: 280px;
-        color: $SECONDARY;
-    }
-
-    button {
-        font-size: 16px;
-        color: $PRIMARY;
-        border-radius: 5px;
-        background: $SECONDARY;
-        height: 30px;
-        width: 80px;
-        z-index: 99;
-        outline: none;
-        margin-top: 30px;
-        padding: 0;
-    }
-
-    #close {
-        background: transparent;
-        font-size: 20px;
-        font-weight: bold;
-        color: $SECONDARY;
-        border: none;
-        width: auto;
-        height: auto;
-        margin: 0;
-        position: relative;
-        top: -5vh;
-        left: 21vh;
-    }
-
-    h3 {
-        margin-top: 0;
-    }
-}
