--- src/components/InfoAndCompare/Info/index.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import Page from '../../Page';
-
-const Info = () => {
-    return (
-        <Page background={'#9BD6DC'}>
-            <div id="Info">
-                <p id="infobox">We don&apos;t sell your data! We just let you compare stuff. Big yay. Swipe to continue.</p>
-            </div>
-        </Page>
-    )
-}
-
-export default Info;
