--- src/components/InfoAndCompare/Info/style.scss
@@ -1,14 +1,0 @@
-#Info {
-    z-index: 99;
-    height: 100vh;
-    width: 100vw;
-    display: flex;
-    justify-content: center;
-
-    #infobox {
-        font-size: 24px;
-        margin-top: 25vh;
-        width: 320px;
-        text-align: center;
-    }
-}
