+++ src/components/Login/style.scss
@@ -13,7 +13,7 @@
     flex-direction: column;
     justify-content: space-around;
     align-items: center;
-    height: 45%;
+    height: 40%;
 }
 
 .Login-description {
