+++ src/components/Popup/style.scss
@@ -1,0 +1,85 @@
+.Popup {
+    width: 90%;
+    height: 80%;
+    border-radius: 10px;
+    top: 5%;
+    overflow: auto;
+    background-color: white;
+    position: absolute;
+    z-index: 100;
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 10%;
+    right: 0;
+    left: 0;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+
+    .topHeader{
+        font-family: Roboto;
+        font-size: 7vw;
+        text-align: center;
+        width: 95%;
+    }
+
+    .innerCollapsible{
+        font-family: Roboto;
+        font-size: 5vw;
+        text-align: center;
+        width: 95%;
+    }
+
+    .innerContent{
+        text-decoration: none;
+    }
+    h2 {
+        font-family: Roboto;
+        font-size: 4vw;
+        text-align: left;
+        width: 95%;
+        text-decoration: none;
+    }
+    p {
+        font-family: Roboto;
+        font-size: 3vw;
+        text-align: left;
+        width: 95%;
+    }
+}
+
+.popupInfo {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(.9, .9, .9, .5);
+    z-index: 99;
+}
+
+.Popup-Close-Button {
+    width: 100%;
+    display: flex;
+    justify-content: flex-end;
+
+    button {
+        border: none;
+        margin: 3px;
+        width: 30px;
+        height: 30px;
+    }
+}
+
+.Popup-Button {
+    padding-left: 90%;
+
+    button {
+        background-color: $WHITE;
+        border: none;
+        color: $SECONDARY;
+        font-size: 16px;
+        height: 20px;
+        width: 20px;
+        border-radius: 50%;
+      }
+}
