+++ src/components/Stories/TopArtists/style.scss
@@ -4,74 +4,6 @@
     flex-direction: column;
     width: 100%;
 
-    #PopupBackground {
-        width: 90vw;
-        height: 80vh;
-        overflow: auto;
-        background-color: white;
-        position: absolute;
-        z-index: 100;
-        margin-left: auto;
-        margin-right: auto;
-        margin-top: 10vh;
-        right: 0;
-        left: 0;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-
-        .topHeader{
-            font-family: Roboto;
-            font-size: 7vw;
-            text-align: center;
-            width: 95%;
-        }
-
-        .innerCollapsible{
-            font-family: Roboto;
-            font-size: 5vw;
-            text-align: center;
-            width: 95%;
-        }
-
-        .innerContent{
-            text-decoration: none;
-        }
-        h2 {
-            font-family: Roboto;
-            font-size: 4vw;
-            text-align: left;
-            width: 95%;
-            text-decoration: none;
-        }
-        p {
-            font-family: Roboto;
-            font-size: 3vw;
-            text-align: left;
-            width: 95%;
-        }
-    
-        button {
-            background-color: $WHITE;
-            border: none;
-            color: $SECONDARY;
-            font-size: 16px;
-            height: 20px;
-            width: 20px;
-            border-radius: 50%;
-        }
-    }
-    
-    .popupInfo {
-        position: absolute;
-        width: 100%;
-        height: 100%;
-        background-color: rgba(.9, .9, .9, .5);
-        z-index: 99;
-    }
-    
-
     .TopArtists-Title{
         display: flex;
         flex-direction: column;
@@ -162,29 +94,4 @@
             }
         }
     } 
-
-    .btnDiv{
-        padding-left: 90%;
-
-        button {
-            background-color: $WHITE;
-            border: none;
-            color: $SECONDARY;
-            font-size: 16px;
-            height: 20px;
-            width: 20px;
-            border-radius: 50%;
-          }
-    }
-
-
-    #arrow {
-        position: relative;
-        left: 90%;
-        top: 2%;
-        width: 33px;
-        height: 33px;
-        background-image: url("../images/arrow-drodpdown.svg");
-        background-size: 100% 100%;
-    }
 }
