+++ src/components/Stories/TopGenres/style.scss
@@ -61,26 +61,3 @@
         background-size: cover;
     }
 }
-
-#popup {
-    width: 80vw;
-    height: auto;
-
-    img {
-        height: 100%;
-        width: 100%;
-    }
-}
-
-#popup-background {
-    position: absolute;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(.9, .9, .9, .5);
-    z-index: -1;
-}
