+++ src/components/Stories/Trendex/index.js
@@ -1,52 +1,15 @@
-import React, { useState } from 'react';
+import React from 'react';
 import Page from '../../Page';
+import Popup from '../../Popup';
 import { User } from '../../../types';
 import { string } from 'prop-types';
-import Collapsible from 'react-collapsible';
+
+const PAGE_NAME = "Trendex";
+const PAGE_INFO = "This is the information about Trendex";
 
 const Trendex = ({ user_1, user_2, my_id }) => {
     const { trendex: user1Trendex, display_name: user1Name } = user_1;
     const { trendex: user2Trendex, display_name: user2Name } = user_2;
-
-    const [showPopup, setShowPopup] = useState(false);
-
-    const renderPopup = () => {
-        return (
-            <div id="PopupBackground">
-                <button onClick={() => setShowPopup(!showPopup)}>X</button>                
-                <Collapsible className="topHeader" openedClassName="topHeader" trigger="About Top Artists" open="true">
-                    <h2> What am I looking at?</h2>
-                    <p>This all has to get filled in- maybe other questions headers too idk- doesn’t always have to be “what am i looking at?</p>
-                </Collapsible>
-                <Collapsible className="topHeader" openedClassName="topHeader" trigger="About Duet" open="true">
-                    <Collapsible className='innerCollapsible' openedClassName='innerCollapsible' trigger="What information does Duet have?">
-                        <p>Duet only stores information given to us from Spotify: your username and your top artists/songs- that’s it! As much as we would like your credit card information or email and password- Spotify’s authentication process prevents us. Dang!</p>
-                    </Collapsible>
-                    <Collapsible className='innerCollapsible' openedClassName='innerCollapsible' trigger="How does Duet work?">
-                        <p>Once you’ve logged in through Spotify’s website and selected a person or friend to compare with, Duet receives the top fifty artists/songs from the past six months of yourself and whoever you chose to compare with. From there, using Spotify’s public api, we look at all the metadata of your music to show how it contrasts! We can even generate playlists to blend your two music tastes like a tropical smoothie on a hot day. Neato!</p>
-                    </Collapsible>
-                    <Collapsible className='innerCollapsible' openedClassName='innerCollapsible' trigger="Who made Duet?">
-                        <p>Duet was made by group of collegiate seniors as a capstone project. We would have been something that cures global warming (easy) or systemic racism and misogyny (child’s play), but unfortunately we were bound by an ancient form of magic known as onomancy, the arcane school of names, to Xolgoth- Daedric prince of the seventh layer of hell/Duet marketing lead. </p>
-                    </Collapsible>
-                    <Collapsible className='innerCollapsible' openedClassName='innerCollapsible' trigger="Why is it mobile only?">
-                        <p>Listen here you little shit...</p>
-                    </Collapsible>
-                    <Collapsible className='innerCollapsible' openedClassName='innerCollapsible' trigger="Contact us">
-                        <p>Have any more questions? You can send us emails at duetwastaken@gmail.com.</p>
-                    </Collapsible>
-                    <Collapsible className='innerCollapsible' openedClassName='innerCollapsible' trigger="Donate!">
-                        <p>To keep Duet ad free, we have been paying for hosting out of pocket :((  consider donating to my venmo: @nicBerg... jk.... unless?</p>
-                    </Collapsible>
-                </Collapsible>
-            </div>
-        )
-    }
-
-    const handleInfoClick = () => {
-        if (showPopup) {
-            setShowPopup(!showPopup);
-        } 
-    }
 
     return (
         <Page background={'#212034'} numPages={5} pageNum={1}>
@@ -85,18 +48,9 @@ const Trendex = ({ user_1, user_2, my_id }) => {
                         <h3 className="Trendex-Score-Label">{user2Name}&apos;s trendex</h3>}
                     {user_2.id == my_id && 
                         <h3 className="Trendex-Score-Label">ur trendex</h3>}
-
                 </div>
-                <div className="btnDiv">
-                    <button onClick={() => setShowPopup(!showPopup)}>?</button>
-                </div>
-
-                { showPopup &&
-                    <div className="popupInfo">
-                        {renderPopup()}
-                    </div>
-                }
             </div>
+            <Popup pageInfo={PAGE_INFO} pageName={PAGE_NAME}/>
         </Page>
     )
 }
