+++ src/styles/styles.scss
@@ -8,9 +8,10 @@
 @import '../components/Stories/Trendex/style.scss';
 @import '../components/Stories/Members/style.scss';
 @import '../components/Stories/Share/style.scss';
-@import '../components/InfoAndCompare/Compare/style.scss';
-@import '../components/InfoAndCompare/Info/style.scss';
+@import '../components/Compare/style.scss';
 @import '../components/Page/style.scss';
 @import '../components/Login/style.scss';
 @import '../components/Stories/Blender/style.scss';
 @import '../components/Button/style.scss';
+@import '../components/Switch/style.scss';
+@import '../components/Popup/style.scss';
